Initial changes to planetPred
Starting to make the initial changes to planetPred which involve changing the code to reflect the column names on our new main.csv file. Currently stuck in an error while running the program on line 206 of simulation_final_nrh.py where two lists are not matching against each other in size.