Bug in simulation_hype_plots
If line 94 is changed to predicted.star_name the code errors out indicating that ClassyReader() does not have a star_name column.

If it's set as predicted.HIP then a different error occurs.

TypeError: unsupported operand type(s) for -: 'str' and 'float' due to the C abundance for the corresponding star being blank.

Additionally noticed that this process is skipping the stars that are not under the HIP catalogue.