Major Bug Fix (get_exo.py)
Fixed a bug which was writing out the incorrect planet data to the main.csv file. The error was due to not updating the planet data dictionary to correlate to the information of the planet of the highest mass (or radius) in the system after the analysis was made. This resulted in the code writing out the final planet for a given star regardless of if it had the highest (mass or radius).