Bug Encountered (simulation_final_nrh.py)
Added code changes to the main simulation files so that planet radius can be utilized rather than the planets mass. Bug is encountered on line 131 (when no golden set is used) or line 67 (when golden set is used). 'ValueError: a must be greater than 0 unless no samples are taken.' Currently trying to figure out the cause.